Function to get typsupport struct from C++ typesupport
