+++ test_msgs/include/test_msgs/action/fibonacci.hpp
@@ -18,80 +18,10 @@
 #include <action_msgs/msg/goal_info.hpp>
 #include <action_msgs/msg/goal_status_array.hpp>
 #include <action_msgs/srv/cancel_goal.hpp>
-#include "rosidl_generator_c/message_type_support_struct.h"
-#include "rosidl_generator_c/service_type_support_struct.h"
+#include "rosidl_generator_c/action_type_support_struct.h"
 #include <test_msgs/action/fibonacci__goal.hpp>
 #include <test_msgs/action/fibonacci__result.hpp>
 #include <test_msgs/action/fibonacci__feedback.hpp>
-
-// TODO(sloretz) define these in rosidl_typesupport_cpp
-namespace rosidl_typesupport_cpp
-{
-  template <typename ACTION>
-  const rosidl_service_type_support_t *
-  get_action_goal_request_typesupport_handle_function(
-    const rosidl_service_type_support_t * handle, const char * identifier)
-  {
-    return get_service_typesupport_handle_function<ACTION::GoalRequestService>(
-        handle, identifier);
-  }
-
-  template <typename ACTION>
-  const rosidl_service_type_support_t *
-  get_action_goal_result_typesupport_handle_function(
-    const rosidl_service_type_support_t * handle, const char * identifier)
-  {
-    return get_service_typesupport_handle_function<ACTION::GoalResponseService>(
-        handle, identifier);
-  }
-
-  template <typename ACTION>
-  const rosidl_service_type_support_t *
-  get_action_cancel_goal_typesupport_handle_function(
-    const rosidl_service_type_support_t * handle, const char * identifier)
-  {
-    return get_service_typesupport_handle_function<ACTION::CancelGoalService>(
-        handle, identifier);
-  }
-
-  template <typename ACTION>
-  const rosidl_message_type_support_t *
-  get_action_feedback_typesupport_handle_function(
-    const rosidl_message_type_support_t * handle, const char * identifier)
-  {
-    return get_message_typesupport_handle_function<ACTION::Feedback>(
-        handle, identifier);
-  }
-
-  template <typename ACTION>
-  const rosidl_message_type_support_t *
-  get_action_status_typesupport_handle_function(
-    const rosidl_message_type_support_t * handle, const char * identifier)
-  {
-    return get_message_typesupport_handle_function<ACTION::GoalStatusMessage>(
-        handle, identifier);
-  }
-}
-// TODO(sloretz) end todo
-
-namespace test_msgs
-{
-namespace action
-{
-struct Fibonacci
-{
-  // Typedefs handy for rcl and rclcpp
-  using CancelGoalService = action_msgs::srv::CancelGoal;
-  using GoalStatusMessage = action_msgs::msg::GoalStatusArray;
-  using GoalRequestService = test_msgs::action::Fibonacci_Goal;
-  using GoalResultService = test_msgs::action::Fibonacci_Result;
-
-  // Typedefs handy for user
-  using Goal = GoalRequestService::Request;
-  using Result = GoalResultService::Response;
-  using Feedback = test_msgs::action::Fibonacci_Feedback;
-};
-}
-}
+#include <test_msgs/action/fibonacci__struct.hpp>
 
 #endif  // TEST_MSGS__ACTION__FIBONACCI_HPP_
